Implemented option to view all saved charts
Main menu option #4 uses try/except block to open and read Portuguese_Verb_Charts.txt in as string, then separates string by line breaks and prints each individual line to user.
If user has already saved verb charts, this option will print out all saved data in the file to the user. If not, an exception is raised and user is asked to check for the file's existence + location (should be in same folder as program).